--login user and merge container PATH
Running a container_env is essentially almost like starting a new terminal, so
we should --login in order to ensure that the PATH is setup correctly.

We also merge the user's PATH with the container's PATH to ensure that the PATH
setup by the container is made use of.

On macOS we do not mount/run the host user profile since I'm assuming its probably
not going to work well on linux containers.